# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Alpine Ecommerce API",
    "version": "1.0.0",
    "description": "# Alpine Ecommerce Backend API\n\nComplete API documentation for Alpine Ecommerce e-commerce platform.\n\n## Features\n- 🔐 JWT Authentication\n- 📦 Product Management\n- 🛒 Order Processing\n- 💳 Payment Integration (Razorpay)\n- 👥 Customer Management\n- 📊 Dashboard & Reports\n- 📝 Error Logging\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-token>\n```\n\n## Base URL\n- **Current Server**: http://13.235.53.15:8081/api/v1\n- **Configuration**: \n  - Set `API_BASE_URL` in .env for full control (recommended)\n  - Or set `PRODUCTION_URL` for production domain\n  - Or set `DOMAIN`, `PORT`, and `PROTOCOL` for auto-construction\n\n## Response Format\nAll responses follow a consistent format:\n```json\n{\n  \"responseStatusCode\": 200,\n  \"isError\": false,\n  \"errorData\": null,\n  \"successData\": { ... }\n}\n```\n\n## Error Handling\nErrors are returned with appropriate status codes:\n```json\n{\n  \"responseStatusCode\": 400,\n  \"isError\": true,\n  \"errorData\": {\n    \"displayMessage\": \"Error description\",\n    \"apiErrorType\": 1\n  }\n}\n```",
    "contact": {
      "name": "Alpine Ecommerce API Support",
      "email": "support@wildvalleyfoods.in"
    },
    "license": {
      "name": "Proprietary"
    },
    "servers": [
      {
...
      }
    ],
...
        }
...
  },
  "paths": {
...
    "/api/v1/invoice/order/{orderId}": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "Get invoice by order ID (Admin only)",
        "description": "Returns invoice for a specific order",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Order ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/order": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create a new order",
        "description": "Creates a new order with payment details",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reqData"
                ],
                "properties": {
                  "reqData": {
                    "type": "object",
                    "required": [
                      "customerId",
                      "items",
                      "totalAmount"
                    ],
                    "properties": {
                      "customerId": {
                        "type": "integer",
                        "example": 1
                      },
                      "items": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "productId": {
                              "type": "integer"
                            },
                            "quantity": {
                              "type": "integer"
                            },
                            "price": {
                              "type": "number"
                            }
                          }
                        }
                      },
                      "totalAmount": {
                        "type": "number",
                        "example": 999.99
                      },
                      "paymentMethod": {
                        "type": "string",
                        "example": "razorpay"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get all orders",
        "description": "Returns all orders (may require authentication)",
        "security": [],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/order/verify": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Verify payment",
        "description": "Verifies Razorpay payment signature and updates order status",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reqData"
                ],
                "properties": {
                  "reqData": {
                    "type": "object",
                    "required": [
                      "orderId",
                      "paymentId",
                      "signature"
                    ],
                    "properties": {
                      "orderId": {
                        "type": "string",
                        "example": "order_123456"
                      },
                      "paymentId": {
                        "type": "string",
                        "example": "pay_123456"
                      },
                      "signature": {
                        "type": "string",
                        "example": "signature_hash"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/order/customer/{customerId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get orders by customer ID",
        "description": "Returns all orders for a specific customer",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/order/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order by ID",
        "description": "Returns a specific order by its ID",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Order ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/order/{id}/status": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Update order status (Admin only)",
        "description": "Updates the status of an order",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Order ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reqData"
                ],
                "properties": {
                  "reqData": {
                    "type": "object",
                    "required": [
                      "status"
                    ],
                    "properties": {
                      "status": {
                        "type": "string",
                        "enum": [
                          "pending",
                          "processing",
                          "shipped",
                          "delivered",
                          "cancelled"
                        ],
                        "example": "shipped"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /register or /login endpoint. Enter your token (with or without \"Bearer \" prefix)"
      }
    }
  },
...
}